LayoutTests/media/audio-playback-restriction-removed-muted-expected.txt

Test that when RequireUserGestureForAudioRateChange is set, setting muted with a user gesture clears the restriction.


RUN(internals.setMediaElementRestrictions(video, "RequireUserGestureForAudioRateChange"))
RUN(video.src = findMediaFile("audio", "content/test"))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
RUN(video.muted = false)
END OF TEST
